Split management session refresh logic
